backend/pet_project/booking/views/booking_view.py
from rest_framework import generics
from booking.models.bookings import Bookings
from booking.serializers.booking_serializer import BookingSerializer

class BookingListCreateView(generics.ListCreateAPIView):
    queryset = Bookings.objects.all()
    serializer_class = BookingSerializer

class BookingDetailView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Bookings.objects.all()
    serializer_class = BookingSerializer
