backend/requirements.txt
django>=4.2
djangorestframework
mysqlclient
